.claude/skills/gemini-imagegen/pyproject.toml
[project]
name = "gemini-imagegen"
version = "1.0.0"
description = "Gemini Image Generation Skill for Claude Code"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
]

[tool.uv]
dev-dependencies = []
